Move IO devices to internal addon
