+++ src/sic/sim/vm/Devices.java
@@ -72,10 +72,7 @@ public class Devices {
     public Devices(int count) {
         assert count > 2;
         devices = new Device[count];
-        setDevice(SICXE.DEVICE_STDIN, new InputDevice(System.in));
-        setDevice(SICXE.DEVICE_STDOUT, new OutputDevice(System.out));
-        setDevice(SICXE.DEVICE_STDERR, new OutputDevice(System.err));
-        for (int i = SICXE.DEVICE_FREE; i < count; i++)
+        for (int i = 0; i < count; i++)
             setDevice(i, new FileDevice(Conversion.byteToHex(i) + ".dev"));
     }
